+++ CHANGELOG.md
@@ -33,6 +33,7 @@ For a full diff see [`88cf670...main`][88cf670...main].
 ### Changed
 
 - Added `EntryList::sortedByPullRequestAscending()` ([#294]), by [@localheinz]
+- Added `ReleaseList::sortedByTagAscending()` ([#295]), by [@localheinz]
 
 [88cf670...main]: https://github.com/ergebnis/keep-a-changelog/compare/88cf670...main
 
@@ -56,5 +57,6 @@ For a full diff see [`88cf670...main`][88cf670...main].
 [#292]: https://github.com/ergebnis/keep-a-changelog/pull/292
 [#293]: https://github.com/ergebnis/keep-a-changelog/pull/293
 [#294]: https://github.com/ergebnis/keep-a-changelog/pull/294
+[#295]: https://github.com/ergebnis/keep-a-changelog/pull/295
 
 [@localheinz]: https://github.com/localheinz